pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobcrew"
version = "1.0.0"
description = "Your AI-Powered Job Search Assistant"
authors = [{name = "kutmur"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
keywords = ["ai", "job-search", "crewai", "agents", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "crewai>=0.28.0",
    "crewai-tools>=0.1.6",
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
]

[project.scripts]
jobcrew = "jobcrew.main:main"

[project.urls]
Homepage = "https://github.com/kutmur/JobCrew"
Repository = "https://github.com/kutmur/JobCrew"
Issues = "https://github.com/kutmur/JobCrew/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["jobcrew*"]
